feat: refactor mcp server - routes + server
